requirements.txt
# Gacha Game Monitor - Dependencies
# 安裝指令: pip install -r requirements.txt

# UI Framework
flet>=0.25.0

# System Tray
pystray>=0.19.0
Pillow>=10.0.0

# System Notifications
plyer>=2.1.0

# Process Monitoring (Phase 3)
psutil>=5.9.0

# Type Hints Support
typing-extensions>=4.0.0
